break up map components
